ham_radio_app/requirements.txt
Flask
Flask-Cors
pymongo
python-dotenv
werkzeug
PyJWT